requirements.txt
praw>=7.6.0
requests>=2.26.0
python-dotenv>=0.19.0
mysql.connector

streamlit

apache-airflow

mysqlclient
SQLAlchemy